For review
Professor Just need your guidance on how to move this forward. I am completely new to this and have not background in coding. Do you have an example that I can use to follow? Also the template you gave us to put in Power point format. Do you want me to put the code in that to? How do I fill it out? Thank you! sorry for any Inconvenience.